static-build: build ovmf with packaged nasm
In order to build ovmf we have built one of its dependencies (nasm) from
source. It happens that https://www.nasm.us/pub/nasm/releasebuilds/2.15.05/nasm-2.15.05.tar.gz
is currently down. This updated the builder image to ubuntu:22.04 that
provides nasm 2.15 packaged so to circumvent problems to download its
sources. Ubuntu 22.04 doesn't not provide the python package, so it was
adjusted as well.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>