gha: refactor install_kubectl
Replace the boilerplate to install kubectl with `install_kubectl()` that
gets a version and optionally a destination directory. The existing
`install_kubectl()` is renamed to `install_kubectl_aks()` because it
actually uses `az` to install the binary matching an AKS installation.

Note: if `kubectl` exists on PATH then it won't be installed again.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>